% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kanji_conversion_table.R
\docType{data}
\name{kanji_conversion_table}
\alias{kanji_conversion_table}
\title{kanji_conversion_table}
\format{A data frame with 2416 rows and 2 variables:
\describe{
  \item{new}{modern, standard form of character}
  \item{old}{older or variant from}
  ...
}}
\source{
\url{http://wwwap.hi.u-tokyo.ac.jp/ships/itaiji_list.jsp}
}
\usage{
kanji_conversion_table
}
\description{
A dataset containing old characters and their modern counterparts. Adapted from The conversion table was adapted from # http://wwwap.hi.u-tokyo.ac.jp/ships/itaiji_list.jsp 確認日：2017年10月29日
}
\keyword{datasets}
