% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Kanji_counter.R
\name{kanji_counter}
\alias{kanji_counter}
\title{Kanji and kana counter}
\usage{
kanji_counter(doc)
}
\arguments{
\item{doc}{A string literal or object}
}
\description{
This function returns a dataframe with statistics for the types of characters in a string or character vector
}
\examples{
kanji_counter("x")
}
\keyword{Japanese,}
\keyword{hiragana,}
\keyword{kanji,}
\keyword{katakana}
