% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NDL_counter.R
\name{NDL_counter}
\alias{NDL_counter}
\title{NDL scraper}
\usage{
NDL_counter(pattern, start_date, end_date)
}
\arguments{
\item{pattern}{a string literal or object}

\item{start_date}{a number for the start year}

\item{end_date}{a number for year the end year}
}
\description{
This function scrapes search results from the National Diet Library's digital collection, counts numbers of hits per years, and generates a full times series with 0 values for missing years.
The search url is http://dl.ndl.go.jp/
}
\examples{
NDL_counter("x", 1920, 1922)
}
\keyword{Japanese,}
\keyword{NDL,}
\keyword{search}
