% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{normalize_kanji}
\alias{normalize_kanji}
\title{Kanji normalizing function}
\usage{
normalize_kanji(doc)
}
\arguments{
\item{doc}{A string literal or object}
}
\description{
This function converts old 旧漢字 and variant 異体字 Japanese characters to their modern equivalents.
The conversion table was adapted from # http://wwwap.hi.u-tokyo.ac.jp/ships/itaiji_list.jsp 確認日：2017年10月29日
}
\examples{
normalize_kanji("亞啞惡蘆鰺壓菴桉")
}
\keyword{Japanese,}
\keyword{kanji}
