% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normalize.R
\name{normalize_kanji}
\alias{normalize_kanji}
\title{Kanji normalizing function}
\usage{
normalize_kanji(doc)
}
\arguments{
\item{doc}{A string literal or object}
}
\description{
This function converts old and variant Japanese characters to their modern equivalents.
}
\examples{
normalize_kanji("x")
}
\keyword{Japanese,}
\keyword{kanji}
